+++ components/atom/tag/src/Actionable/settings.js
@@ -3,7 +3,7 @@ import cx from 'classnames'
 export const RIGHT_ICON_PLACEMENT = 'right'
 export const LEFT_ICON_PLACEMENT = 'left'
 
-export const getClassNames = function({className, disabled}) {
+export const getClassNames = function ({className, disabled}) {
   return cx(
     'sui-AtomTag-actionable',
     disabled && 'sui-AtomTag--disabled',