+++ components/atom/validationText/src/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 
 import {TYPES, getClassNames} from './settings.js'
 
-const AtomValidationText = function({type, text}) {
+const AtomValidationText = function ({type, text}) {
   return <span className={getClassNames(type)}>{text}</span>
 }