+++ components/molecule/checkboxField/demo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+# 1.2.0 (2022-02-09)
+
+
+### Bug Fixes
+
+* **components/molecule/checkboxField:** add checkboxField disabled ([2b1de65](https://github.com/SUI-Components/sui-components/commit/2b1de6503146bf10d88f7480136fd472a12741f6))
+
+
+### Features
+
+* **components/molecule/checkboxField/demo:** Add articles for checkboxField ([ce60d7b](https://github.com/SUI-Components/sui-components/commit/ce60d7b0bc11663dc35fb0639c52be6c2753e6db))
+* **components/molecule/checkboxField/demo:** Add classname ([d19e149](https://github.com/SUI-Components/sui-components/commit/d19e149231a218517c7d990e79b1c79207443fde))
+* **components/molecule/checkboxField/demo:** minor fix ([5e0232b](https://github.com/SUI-Components/sui-components/commit/5e0232b73c662f3516ec34eb38c312971da3cf87))
+* **components/molecule/checkboxField/demo:** remove className ([7c6f327](https://github.com/SUI-Components/sui-components/commit/7c6f32783c980ed88a34aafaf78ffcff4879d483))
+
+
+
 # 1.1.0 (2021-10-13)